refac: rename list to s
